Use `restore-state-none` feature in optional dependency `critical-section` (#371)
This should provide a slightly user-friendlier error message in the case that a user enables a different `restore-state` in `critical-section`.

Right now it's possible to e.g. enable the `std` feature on `critical-section`; `cargo` will not complain but `rustc` will complain about the return types of the `CriticalSection::enter` implementation.