Improve delay/timer errors and rounding calculations (#374)
* delay: Remove unnecessary warning suppressions

* Improve system tick division and rounding

TICK_PERIOD_MS might include an error, if 1000 ms is not divisible by the configured tick rate.
Do not use it internally for converting between ms and ticks.

Also, always round up when dividing.

* Round delay times up

It's typically desired to round delay times up instead of down,
because a delay usually waits for a thing (e.g. hardware) to finish.
If we would round down, it is more likely to cause ill effects than
rounding up. The user code usually expects delaying a bit more.

In case of the coarse timer the "waiting a bit more" may be fairly
large. But if the user code uses a coarse timer for small delays
it's a problem in the user code.

This change also converts a couple of 'as' conversions to 'into()' calls.

* Export tick rate in Hz instead of tick period

because the division may introduce a rounding error on certain configurations.

* Make TickType::new_millis() pub

as counterpart to the existing as_millis()

* Ceil round Duration when converting to ms

We carefully saturate, even though it's unlikely that big durations
that overflow an u64 ms are supplied.

* delay: Replace constants with names

* Implement DelayNs::delay_ms()

The default implementation from embedded_hal is not ideal for
us, because it calls delay_ns() which we don't really support.
It causes unnecessary computation and rounding.